chore: move xcm bridge hook to correct directory
